refactor(errors): split worktree-remote-error helpers to a dedicated module
git-remote-error.ts was mixing two responsibilities: general git
remote error helpers (normalizeGitErrorMessage, isNoUpstreamError,
stripCredentialsFromMessage, formatSubmodulePushFailureDetail) and
worktree-specific classifier+formatter+parser (classify/format/parse
RefreshBaseRefErrorCode). Split now so the next domain-specific
classifier doesn't pile on.

- New: src/shared/worktree-remote-error.ts (worktree-specific)
- Kept: src/shared/git-remote-error.ts (general helpers, kept in place)
- 3 consumers updated to import from the new module
- Test file split: 20 tests in git-remote-error, 11 in worktree-remote-error
